// modules/vectors/src/vec4.d.ts
import { mat4 } from './mat4';
export declare class vec4 extends Float32Array {
    static readonly zero: Readonly<vec4>;
    static readonly one: Readonly<vec4>;
    constructor(values?: number[]);
    get x(): number;
    set x(x: number);
    get y(): number;
    set y(y: number);
    get z(): number;
    set z(z: number);
    get w(): number;
    set w(w: number);
    get xyzw(): number[];
    set xyzw(xyzw: number[]);
    get length(): number;
    get squaredLength(): number;
    reset(): vec4;
    copy(dest?: null | vec4): vec4;
    negate(dest?: null | vec4): vec4;
    equals(vector: vec4, threshold?: number): boolean;
    add(vector: vec4, dest?: null | vec4): vec4;
    subtract(vector: vec4, dest?: null | vec4): vec4;
    multiply(vector: vec4, dest?: null | vec4): vec4;
    divide(vector: vec4, dest?: null | vec4): vec4;
    scale(scalar: number, dest?: null | vec4): vec4;
    normalize(dest?: null | vec4): vec4;
    transform(matrix: mat4, dest?: null | vec4): vec4;
    serialize(): number[];
    static deserialize(values: number[]): vec4;
    static absolute(vector: vec4, dest?: null | vec4): vec4;
    static minimum(vector: vec4, vector2: vec4, dest?: null | vec4): vec4;
    static maximum(vector: vec4, vector2: vec4, dest?: null | vec4): vec4;
    static mix(vector: vec4, vector2: vec4, time: number, dest?: null | vec4): vec4;
    static add(vector: vec4, vector2: vec4, dest?: null | vec4): vec4;
    static subtract(vector: vec4, vector2: vec4, dest?: null | vec4): vec4;
    static multiply(vector: vec4, vector2: vec4, dest?: null | vec4): vec4;
    static divide(vector: vec4, vector2: vec4, dest?: null | vec4): vec4;
    static scale(vector: vec4, scalar: number, dest?: null | vec4): vec4;
    static normalize(vector: vec4, dest?: null | vec4): vec4;
    static sum(...vectors: vec4[]): vec4;
    static difference(...vectors: vec4[]): vec4;
    static product(...vectors: vec4[]): vec4;
    static division(...vectors: vec4[]): vec4;
}